# app/static/swagger.yaml
swagger: "2.0"
info:
  title: Emotion Tracker API
  description: >
    The Emotion Tracker API allows users to register, log in, and manage their emotional entries. Each user can create mood entries that include one of three predefined moods, a list of custom feelings (created on the frontend), an optional diary entry, sleep duration, and a date. The API supports creating, reading, updating, and deleting users and mood entries. Feelings are user-generated and saved per entry through a many-to-many relationship. Authentication and user management are supported, and all data is returned in JSON format. This API is designed to integrate smoothly with frontend applications for tracking mental and emotional health.
  version: "1.0.0"
host: 127.0.0.1:5000
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

paths:
  /users/login:
    post:
      tags:
        - users
      summary: Login endpoint for users, generates a jwt token
      description: Validates user credentials, generates and returns a user specific token
      parameters:
        - in: body
          name: body
          description: Login Credentials
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: Login successful
          schema:
            $ref: "#/definitions/LoginResponse"
        400:
          description: Missing email or password
          schema:
            type: object
            properties:
              error:
                type: string
                example: Email and password are required
        401:
          description: Invalid email or password
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid credentials

  /users:
    post:
      tags:
        - users
      summary: Endpoint to create a user
      description: >
        Creates a new user to be able to track their moods (name, email, and password are required fields). is_admin field is defaulted to false, and image has a default user image.
      required: true
      schema:
        $ref: "#/definitions/CreateUserRequest"
      parameters:
        - in: body
          name: body
          description: Create a user
          required: true
          schema:
            $ref: "#/definitions/CreateUserResponse"
      responses:
        201:
          description: User created successfully
          schema:
            $ref: "#/definitions/CreateUserResponse"
        400:
          description: Validation error (e.g. missing or invalid fields)
          schema:
            type: object
            example:
              email:
                - "Not a valid email address"
              password:
                - "Missing data for required field"
              name:
                - "Shorter than minimum length 3"
        409:
          description: Email already registered
          schema:
            type: object
            example:
              error: "Email already registered"

    get:
      tags:
        - users
      summary: Get all users
      description: Endpoint to retrieve a list of all users
      responses:
        200:
          description: A list of users
          schema:
            type: array
            items:
              $ref: "#/definitions/UserResponse"
          examples:
            application/json:
              - id: 1
                name: "John Doe"
                email: "john@example.com"
                is_admin: false
                image: "https://example.com/profile1.jpg"
              - id: 2
                name: "Jane Smith"
                email: "jane@example.com"
                is_admin: true
                image: "https://example.com/profile2.jpg"
        404:
          description: No users found
          schema:
            type: object
            example:
              message: "No users found"

    patch:
      tags:
        - users
      summary: Update a user's information
      description: >
        Allows partial updates to a user's account. Any combination of fields can be updated. Passwords will be hashed before being stored. Returns the updated user data upon success. A user must be logged in to access this route.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Fields to update for the user
          schema:
            $ref: "#/definitions/UserUpdateRequest"
      responses:
        200:
          description: User updated successfully
          schema:
            $ref: "#/definitions/UserResponse"
        400:
          description: Bad request - invalid data
          schema:
            $ref: "#/definitions/ValidationError"
        404:
          description: User not found
          schema:
            type: object
            example:
              error: "User not found"
        409:
          description: Email already in use
          schema:
            type: object
            example:
              error: "Email already in use"
        500:
          description: Database error
          schema:
            type: object
            example:
              error: "Database error"

    delete:
      tags:
        - users
      summary: Delete a user
      description: Endpooint to delete a logged in user from the database.
      security:
        - Bearer: []
      responses:
        200:
          description: User deleted successfully
          schema:
            type: object
            properties:
              message:
                type: string
                example: "User 5 deleted successfully"
        404:
          description: User not found
          schema:
            type: object
            properties:
              message:
                type: string
                example: "User not found"

  /users/{user_id}:
    get:
      tags:
        - users
      summary: Get a single user by ID
      description: Retrieves a single user object based on the provided user ID.
      parameters:
        - in: path
          name: user_id
          required: true
          type: integer
          description: ID of the user to retrieve
      responses:
        200:
          description: user found and returned successfully
          schema:
            $ref: "#/definitions/UserResponse"
        404:
          description: No user found with that ID
          schema:
            type: object
            properties:
              message:
                type: string
                example: "No user found with that id"

  /feelings:
    post:
      tags:
        - feelings
      summary: Create a new user feeling
      description: >
        Creates a new feeling for a specific mood, requires a mood_name. You can either set it up for users to create the feelings, or alternatively set predefined feelings for each predefined mood.
      parameters:
        - in: body
          name: body
          description: Feeling data to create a new feeling
          required: true
          schema:
            $ref: "#/definitions/CreateFeelingRequest"
      responses:
        201:
          description: Feeling created successfully
          schema:
            $ref: "#/definitions/CreateFeelingResponse"
        400:
          description: Validation error (e.g. missing or invalid fields)
          schema:
            $ref: "#/definitions/FeelingsValidationError"

  /usermoods:
    post:
      tags:
        - usermoods
      summary: Creates a new user mood entry
      description: >
        Creates a new mood entry for a user with a diary entry, sleep time, and associated feelings. Requires user ID, mood ID, sleep time, and a list of at least one feeling IDs. Mood ID's are predefined and are as follows: {"id": 1, mood_type: "happy"}, {"id": 2, "mood_type": "neutral"}, {"id": 3, "moode_type": "sad"}.
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: User mood data to create
          required: true
          schema:
            $ref: "#/definitions/UserMoodRequest"
      responses:
        201:
          description: UserMood created successfully
          schema:
            $ref: "#/definitions/UserMoodResponse"
        400:
          description: Validation error (e.g. missing or invalid fields)
          schema:
            type: object
            example:
              feelings:
                - "One or more feeling IDs are invalid"
              mood:
                - "Missing data for required field"
        401:
          description: Unauthorized - missing or invalid token
          schema:
            type: object
            example:
              message: "Token is missing or invalid"
        403:
          description: Forbidden - user not authorized to delete this route
          schema:
            type: object
            example:
              message: "Token is missing or invalid"
        500:
          description: Server error
          schema:
            type: object
            example:
              error: "Internal server error"

    get:
      tags:
        - usermoods
      summary: Get a list of all usermood entries for a user
      description: Endpoint to retrieve a list of all usermood entries for a logged in user
      security:
        - Bearer: []
      parameters:
        - in: path
          name: usermood_id
          required: true
          type: integer
          description: ID of usermood to delete
      responses:
        200:
          description: Usermoods retrieved successfully
          schema:
            $ref: "#/definitions/UserMoodResponse"
        401:
          description: Unauthorized - missing or invalid token
          schema:
            type: object
            example:
              message: "Token is missing or invalid"
        403:
          description: Forbidden - user not authorized to delete this route
          schema:
            type: object
            example:
              message: "Token is missing or invalid"
        404:
          description: User not found
          schema:
            type: object
            properties:
              message:
                type: string
                example: "No usermood entries found"

    delete:
      tags:
        - usermoods
      summary: Delete a specific UserMood entry
      description: >
        Deletes a UserMood entry by ID if it belongs to the authenticated user.
        Requires a valid JWT token for authentication.
      security:
        - Bearer: []
      parameters:
        - in: path
          name: path
          description: ID of UserMood to delete
          required: true
          type: integer
      responses:
        200:
          description: UserMood deleted successfully
          schema:
            type: object
            example:
              message: "UserMood deleted successfully"
        401:
          description: Unauthorized - missing or invalid token
          schema:
            type: object
            example:
              message: "Token is missing or invalid"
        403:
          description: Forbidden - user not authorized to delete this route
          schema:
            type: object
            example:
              message: "Token is missing or invalid"
        404:
          description: UserMood not found
          schema:
            type: object
            example:
              message: "UserMood not found"

definitions:
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"

    required:
      - email
      - password

  LoginResponse:
    type: object
    properties:
      message:
        type: string
        example: "Login successful"
      token:
        type: string
        example: "nlkKNNniNIKNJALSKll..."
      user:
        type: object
        properties:
          id:
            type: integer
            example: 1
          name:
            type: string
            example: "John Doe"
          email:
            type: string
            example: "john@example.com"
          is_admin:
            type: boolean
            example: false

  CreateUserRequest:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      password:
        type: string
      is_admin:
        type: boolean
      image:
        type: string

    required:
      - name
      - email
      - password

  CreateUserResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "john@example.com"
      is_admin:
        type: boolean
        example: false

  UserResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "john@example.com"
      is_admin:
        type: boolean
        example: false
      image:
        type: string
        example: "https://example.com/profile.jpg"

  UserUpdateRequest:
    type: object
    properties:
      name:
        type: string
        example: "Jane Doe"
      email:
        type: string
        example: "jane@example.com"
      password:
        type: string
        example: "newsecurepassword123"
      image:
        type: string
        example: "https://example.com/profile.jpg"
      is_admin:
        type: boolean
        example: false

  CreateFeelingRequest:
    type: object
    properties:
      feeling_name:
        type: string
        example: "Excited"

    required:
      - feeling_name

  CreateFeelingResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      feeling_name:
        type: string
        example: "Excited"

  UserMoodRequest:
    type: object
    properties:
      user:
        type: integer
        example: 2
      mood:
        type: integer
        example: 3
      diary_entry:
        type: string
        example: "I'm tired"
      sleep_time:
        type: integer
        example: 7
      feelings:
        type: array
        items:
          type: integer
        example: [3, 4]

  UserMoodResponse:
    type: object
    properties:
      id:
        type: integer
        example: 6
      user:
        type: integer
        example: 1
      mood:
        type: integer
        example: 3
      mood_info:
        $ref: "#/definitions/Mood"
      feelings_info:
        type: array
        items:
          $ref: "#/definitions/Feeling"
      diary_entry:
        type: integer
        example: "I'm tired"
      sleep_time:
        type: integer
        example: 7
      date:
        type: string
        format: date
        example: "2025-06-02"

  Mood:
    type: object
    properties:
      id:
        type: integer
        example: 3
      mood_type:
        type: string
        example: "sad"

  Feeling:
    type: object
    properties:
      id:
        type: integer
        example: 4
      feeling_type:
        type: string
        example: "anxious"

  ValidationError:
    type: object
    description: Returned when required fields are missing or invalid
    example:
      email: "Not a valid email address."
      password: "Missing data for required field."
      name: "Shorter than minimum length 3."

  FeelingsValidationError:
    type: object
    description: Returned when required fields are missing or invalid
    example:
      feeling_name: "Invalid feeling name."

  ConflictError:
    type: object
    description: Returned when trying to register with an email that already exists
    example:
      error: "Email already registered"